refactor: replace custom helpers with standard libraries

+++ test/DynamicFeeManager.t.sol
@@ -19,7 +19,7 @@ event AlreadyInitialized(PoolId indexed id);
 /// versions start using other IPoolPolicy methods, tests will revert. Add the needed methods then.
 contract StubPolicy {
     function getDefaultDynamicFee() external pure returns (uint256) {
-        return 3_000;                // 0.30 % – well below 2**96-1
+        return 3_000; // 0.30 % – well below 2**96-1
     }
     /* everything else can be left un-implemented for this unit-test */
 }
@@ -44,16 +44,16 @@ contract DynamicFeeManagerTest is Test {
         IPoolPolicy policy = IPoolPolicy(address(stub));
 
         oracle = new TruncGeoOracleMulti(
-            dummyPM,                // pool-manager
-            address(this),          // governance
-            policy                  // policy manager
+            dummyPM, // pool-manager
+            address(this), // governance
+            policy // policy manager
         );
 
         // Mock oracle setup
         dfm = new DynamicFeeManager(
-            policy,          // IPoolPolicy
+            policy, // IPoolPolicy
             address(oracle), // oracle
-            address(this)    // authorised hook (this test contract)
+            address(this) // authorised hook (this test contract)
         );
     }
 
@@ -66,14 +66,14 @@ contract DynamicFeeManagerTest is Test {
 
     function testCapMapping() external {
         PoolId pid = PoolId.wrap(bytes32(uint256(1)));
-        
+
         CapTestCase[] memory cases = new CapTestCase[](4);
         cases[0] = CapTestCase(42, 4200, "typical small cap");
         cases[1] = CapTestCase(1000, 100000, "medium cap");
         cases[2] = CapTestCase(16_777_215, 1_677_721_500, "uint24 upper-bound");
         cases[3] = CapTestCase(1, 100, "minimum cap");
 
-        for (uint i; i < cases.length; ++i) {
+        for (uint256 i; i < cases.length; ++i) {
             CapTestCase memory tc = cases[i];
             _setCap(pid, tc.cap);
             assertEq(dfm.baseFeeFromCap(pid), tc.expectPpm, tc.note);
@@ -82,24 +82,24 @@ contract DynamicFeeManagerTest is Test {
 
     function testInitializeIdempotent() public {
         PoolId pid = PoolId.wrap(bytes32(uint256(1)));
-        
+
         // ensure a non-zero cap so the base-fee is > 0
         _setCap(pid, 42);
 
         // First initialization should succeed
         dfm.initialize(pid, 0);
         uint256 initialBaseFee = dfm.baseFeeFromCap(pid);
-        
+
         // Second initialization should not revert and should emit event with correct args
         vm.expectEmit(true, true, false, true);
         emit AlreadyInitialized(pid);
         dfm.initialize(pid, 0);
-        
+
         // Third initialization should behave the same way
         vm.expectEmit(true, true, false, true);
         emit AlreadyInitialized(pid);
         dfm.initialize(pid, 0);
-        
+
         // Verify state remained unchanged throughout
         uint256 finalBaseFee = dfm.baseFeeFromCap(pid);
         assertEq(finalBaseFee, initialBaseFee, "Base fee should remain unchanged after multiple inits");
@@ -108,4 +108,4 @@ contract DynamicFeeManagerTest is Test {
 }
 
 // Legacy step-based tests removed as they no longer apply to the new fee model
-// which derives fees directly from oracle caps (1 tick = 100 ppm = 0.01%) 
+// which derives fees directly from oracle caps (1 tick = 100 ppm = 0.01%)
